refactor: extract to interface
