+++ .gitignore
@@ -1,2 +1,5 @@
+/.idea
+/.phpunit.result.cache
+/docs/
 /phpcs-cache.json
 /vendor/
