--- docs/index.html
@@ -1,143 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
- <head>
-  <meta charset="UTF-8">
-  <title>Code Coverage for /Users/anthony/git/managur/Collection/src</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link href=".css/bootstrap.min.css" rel="stylesheet" type="text/css">
-  <link href=".css/style.css" rel="stylesheet" type="text/css">
-  <link href=".css/custom.css" rel="stylesheet" type="text/css">
-  <!--[if lt IE 9]>
-  <script src=".js/html5shiv.min.js"></script>
-  <script src=".js/respond.min.js"></script>
-  <![endif]-->
- </head>
- <body>
-  <header>
-   <div class="container">
-    <div class="row">
-     <div class="col-md-12">
-      <ol class="breadcrumb">
-        <li class="active">/Users/anthony/git/managur/Collection/src</li>
-        <li>(<a href="dashboard.html">Dashboard</a>)</li>
-
-      </ol>
-     </div>
-    </div>
-   </div>
-  </header>
-  <div class="container">
-   <table class="table table-bordered">
-    <thead>
-     <tr>
-      <td>&nbsp;</td>
-      <td colspan="9"><div align="center"><strong>Code Coverage</strong></div></td>
-     </tr>
-     <tr>
-      <td>&nbsp;</td>
-      <td colspan="3"><div align="center"><strong>Lines</strong></div></td>
-      <td colspan="3"><div align="center"><strong>Functions and Methods</strong></div></td>
-      <td colspan="3"><div align="center"><strong>Classes and Traits</strong></div></td>
-     </tr>
-    </thead>
-    <tbody>
-     <tr>
-      <td class="success">Total</td>
-      <td class="success big">       <div class="progress">
-         <div class="progress-bar progress-bar-success" role="progressbar" aria-valuenow="98.10" aria-valuemin="0" aria-valuemax="100" style="width: 98.10%">
-           <span class="sr-only">98.10% covered (success)</span>
-         </div>
-       </div>
-</td>
-      <td class="success small"><div align="right">98.10%</div></td>
-      <td class="success small"><div align="right">103&nbsp;/&nbsp;105</div></td>
-      <td class="success big">       <div class="progress">
-         <div class="progress-bar progress-bar-success" role="progressbar" aria-valuenow="96.43" aria-valuemin="0" aria-valuemax="100" style="width: 96.43%">
-           <span class="sr-only">96.43% covered (success)</span>
-         </div>
-       </div>
-</td>
-      <td class="success small"><div align="right">96.43%</div></td>
-      <td class="success small"><div align="right">27&nbsp;/&nbsp;28</div></td>
-      <td class="danger big">       <div class="progress">
-         <div class="progress-bar progress-bar-danger" role="progressbar" aria-valuenow="50.00" aria-valuemin="0" aria-valuemax="100" style="width: 50.00%">
-           <span class="sr-only">50.00% covered (danger)</span>
-         </div>
-       </div>
-</td>
-      <td class="danger small"><div align="right">50.00%</div></td>
-      <td class="danger small"><div align="right">1&nbsp;/&nbsp;2</div></td>
-     </tr>
-
-     <tr>
-      <td class="success"><span class="glyphicon glyphicon-file"></span> <a href="Collection.php.html">Collection.php</a></td>
-      <td class="success big">       <div class="progress">
-         <div class="progress-bar progress-bar-success" role="progressbar" aria-valuenow="98.06" aria-valuemin="0" aria-valuemax="100" style="width: 98.06%">
-           <span class="sr-only">98.06% covered (success)</span>
-         </div>
-       </div>
-</td>
-      <td class="success small"><div align="right">98.06%</div></td>
-      <td class="success small"><div align="right">101&nbsp;/&nbsp;103</div></td>
-      <td class="success big">       <div class="progress">
-         <div class="progress-bar progress-bar-success" role="progressbar" aria-valuenow="96.15" aria-valuemin="0" aria-valuemax="100" style="width: 96.15%">
-           <span class="sr-only">96.15% covered (success)</span>
-         </div>
-       </div>
-</td>
-      <td class="success small"><div align="right">96.15%</div></td>
-      <td class="success small"><div align="right">25&nbsp;/&nbsp;26</div></td>
-      <td class="danger big">       <div class="progress">
-         <div class="progress-bar progress-bar-danger" role="progressbar" aria-valuenow="50.00" aria-valuemin="0" aria-valuemax="100" style="width: 50.00%">
-           <span class="sr-only">50.00% covered (danger)</span>
-         </div>
-       </div>
-</td>
-      <td class="danger small"><div align="right">50.00%</div></td>
-      <td class="danger small"><div align="right">1&nbsp;/&nbsp;2</div></td>
-     </tr>
-
-     <tr>
-      <td class="success"><span class="glyphicon glyphicon-file"></span> <a href="functions.php.html">functions.php</a></td>
-      <td class="success big">       <div class="progress">
-         <div class="progress-bar progress-bar-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
-         </div>
-       </div>
-</td>
-      <td class="success small"><div align="right">100.00%</div></td>
-      <td class="success small"><div align="right">2&nbsp;/&nbsp;2</div></td>
-      <td class="success big">       <div class="progress">
-         <div class="progress-bar progress-bar-success" role="progressbar" aria-valuenow="100.00" aria-valuemin="0" aria-valuemax="100" style="width: 100.00%">
-           <span class="sr-only">100.00% covered (success)</span>
-         </div>
-       </div>
-</td>
-      <td class="success small"><div align="right">100.00%</div></td>
-      <td class="success small"><div align="right">2&nbsp;/&nbsp;2</div></td>
-      <td class=" big"></td>
-      <td class=" small"><div align="right">n/a</div></td>
-      <td class=" small"><div align="right">0&nbsp;/&nbsp;0</div></td>
-     </tr>
-
-
-    </tbody>
-   </table>
-   <footer>
-    <hr/>
-    <h4>Legend</h4>
-    <p>
-     <span class="danger"><strong>Low</strong>: 0% to 50%</span>
-     <span class="warning"><strong>Medium</strong>: 50% to 90%</span>
-     <span class="success"><strong>High</strong>: 90% to 100%</span>
-    </p>
-    <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 6.0.8</a> using <a href="https://secure.php.net/" target="_top">PHP 7.1.22</a> with <a href="https://xdebug.org/">Xdebug 2.6.1</a> and <a href="https://phpunit.de/">PHPUnit 7.4.0</a> at Sat Oct 6 16:50:37 UTC 2018.</small>
-    </p>
-   </footer>
-  </div>
-  <script src=".js/jquery.min.js" type="text/javascript"></script>
-  <script src=".js/bootstrap.min.js" type="text/javascript"></script>
-  <script src=".js/holder.min.js" type="text/javascript"></script>
- </body>
-</html>
